requirements.txt
streamlit
python-dotenv
openai
llama-index
llama-index-core
llama-index-llms-openai
llama-index-llms-replicate
llama-index-embeddings-huggingface
nltk
pypdf2
hupper